requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymongo==4.9.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-dotenv==1.0.1
python-multipart==0.0.12
